// src/app/interfaces/point.ts
import  { Position } from './position';

export interface Point {
    nome?: string;
    atendimento?: string;
    id?: string;
    title?: string;
    description?: string;
    biografia?: string;
    servicos?: string;
    position?: Position;
    createdAt?: number;
    telefone?: string;
    whats?: string;
    userId?: string;
    endereco?: string;
    icon?: string;
}